made git repo

+++ mean.py
@@ -3,7 +3,7 @@ def mean(s):
     c = 0
     mean = 0
     while len(sa) > c:
-        sa[c] = int(sa[c])
+        sa[c] = float(sa[c])
         mean = mean + sa[c]
         c += 1
     return mean
